Give up
This is a great story about a young boy who tried to contribute
to a Bengali programming language. He would soon realize that he
made a grave mistake. He tried to implement a `-p` flag to hide
file paths from error messages, but failed miserably.

Moral of the story: NEVER try to hide file paths.

(not leaving the project or anything, just wanted to make a
 dramatic commit message for no reason)